refactor: move ErrorStruct test to moonlink_error
